Convert `experiments.py` into `experiments` package
Split into `experiments/__init__.py`, `experiments/input.py`, `experiments/output.py`, and `experiments/utils.py`.
